Wait for last byte to be sent before continuing
The send_byte function already waits for the USI to become available. So
there is no need to wait in the for loop. However you should wait after all
bytes have been sent to continue safely, as the interrupt makes sure the
timers are given back to the Arduino core.

This example doesn't work right now as the Arduino core tries to delay using
timers still in use by the USI because the interrupt hasn't fired yet.